Extracted Schema into it's own class along with the tests.
[#149550919]
