+++ lib/services/service_brokers/v2.rb
@@ -5,6 +5,7 @@ require 'services/service_brokers/v2/catalog'
 require 'services/service_brokers/v2/catalog_service'
 require 'services/service_brokers/v2/catalog_plan'
 require 'services/service_brokers/v2/catalog_schemas'
+require 'services/service_brokers/v2/schema'
 require 'services/service_brokers/v2/http_client'
 require 'services/service_brokers/v2/client'
 require 'services/service_brokers/v2/orphan_mitigator'
